Verify that $textAttribute["html'] is set before accessing it.
Original code did not do the above check. Here is the info
from the JIRA ticket for this bug:
Here is a sample error from error_log (in between the BEGIN_BLURB>>>>>>
and the <<<<<<END_BLURB):
BEGIN_BLURB>>>>>>
[Tue Jun 16 15:25:53 2015] [error] [client 128.59.154.32] PHP Notice:
Undefined index: html in /cul/cul0/lito/vmounts/handel/var-www/omeka/
versions/v222devCulFromHaydn/application/models/Mixin/ElementText.php
on line 514, referer: https://exhibitions-dev.cul.columbia.edu/admin/
items/edit/10198
<<<<<<END_BLURB
NOTE: though this bug generates an entry in the error log
(/culprod/omeka/logs/error_log), it does not affect the display
of the page.
It seems that the $textAttribute entry is not being set for elements that
have controlled vocabulary via the Simple Vocab plugin. Regardless, the Omeka
code referred to in the error should probably check to make sure that
$textAttribute['html'] is set before trying to access it. And, looking at the
code for Omeka 2.3, this check is now performed in the latest Omeka code.
Since we are not ready yet to upgrade from 2.2.2 to 2.3, I will put in the fix
by hand in the Omeka code running on the server.
Fix is in the file that generates the error:
application/models/Mixin/ElementText.php
To exercise the bug, edit an item, and the error will show up in the log when
you save the item. One error per field controlled via Simple Vocab.